requirements.txt
alembic==1.13.1
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
asyncpg==0.29.0
automium==0.2.6
automium-web==0.1.1
backcall==0.2.0
black==22.12.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.4
executing==1.2.0
filelock==3.6.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.1
jedi==0.18.2
Jinja2==3.0.3
jupyter_client==8.3.0
jupyter_core==5.3.1
legacy==0.1.7
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.0
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.11.0
virtualenv==20.13.3
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.16.0
python-docx
openai
fastapi
uvicorn
python-multipart
